% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mica_methods.R
\name{as.matrix.mica}
\alias{as.matrix.mica}
\title{Extract the Completed Matrix from a MICA Object}
\usage{
\method{as.matrix}{mica}(x, ...)
}
\arguments{
\item{x}{A \code{"mica"} object.}

\item{...}{Unused.}
}
\value{
A completed positive-definite correlation matrix.
}
\description{
Extract the Completed Matrix from a MICA Object
}
