% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{calculate_cdf}
\alias{calculate_cdf}
\title{Calculate empirical CDF}
\usage{
calculate_cdf(data)
}
\description{
Given a set of intensity values for an image, will interpolate the
empirical CDF to a new domain.
}
