% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cdf.R
\name{estimate_cdf}
\alias{estimate_cdf}
\title{create labeled dataframe of vectorized voxel intensities for a list of nifti objects}
\usage{
estimate_cdf(intensity_df, intensity_maximum, rescale_intensities = FALSE,
  ...)
}
\arguments{
\item{intensity_df}{Dataframe of vectorized intensity values.}

\item{intensity_maximum}{Maximum value of intensity for creating grid over which to evaluate CDF and
estimate warping functions. Needs to be chosen based on inspection of data.}

\item{rescale_intensities}{If \code{TRUE}, intensities will be rescaled by their 99.9% quantile.
Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to or from other functions.}
}
\description{
Function used along with \code{mica::vectorize_image()} to vectorize images and place in a
dataframe, where names for each image are drawn from the filepath of the
image.
}
