% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_description.R
\docType{data}
\name{estrogene_data}
\alias{estrogene_data}
\title{Estrogene data list}
\format{
A list of data frames with rows for the genes and columns for the samples.
}
\usage{
estrogene_data
}
\description{
The dataset comes from the EstroGene project, which integrates estrogen-related datasets, including RNA-Seq and microarray data.
This dataset focuses on gene expression studies involving estrogen receptor-positive (ER+) samples treated with estradiol (E2) at doses greater than 1nM. The samples were categorized by cell line, sequencing technology, and treatment duration (short, medium, long).
Data normalization was performed using TMM and ComBat methods. The resulting pooled studies include MCF7 microarray, MCF7 RNA-Seq, and T47D RNA-Seq
}
\references{
Li, Z., Li, T., Yates, M. E., Wu, Y., Ferber, A., Chen, L., ... & Lee, A. V. (2023). The EstroGene database reveals diverse temporal, context-dependent, and bidirectional estrogen receptor regulomes in breast cancer. Cancer Research, 83(16), 2656-2674.
}
\keyword{datasets}
