% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_description.R
\docType{data}
\name{estrogene_label}
\alias{estrogene_label}
\title{Estrogene label list}
\format{
A list of vectors (short, medium, long) for the treatment duration of the samples.
}
\usage{
estrogene_label
}
\description{
A list of vectors for the labels of \link{estrogene_data}.
}
\references{
Li, Z., Li, T., Yates, M. E., Wu, Y., Ferber, A., Chen, L., ... & Lee, A. V. (2023). The EstroGene database reveals diverse temporal, context-dependent, and bidirectional estrogen receptor regulomes in breast cancer. Cancer Research, 83(16), 2656-2674.
}
\keyword{datasets}
