% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_cdf.R
\name{interpolate_cdf}
\alias{interpolate_cdf}
\title{Interpolate CDF}
\usage{
interpolate_cdf(data, intensity_grid, long_to_long = FALSE)
}
\arguments{
\item{data}{dataset with intensity as a variable - this is the set of original intensity values}

\item{intensity_grid}{set of intensities shorter than the length of the original data}

\item{long_to_long}{If TRUE, outputs CDFs for full intensity grid.
If FALSE, takes sparse intensity grid as input and outputs CDF the length of the original data.}
}
\description{
Given a set of intensity values for an image, will interpolate the empirical CDF to a new domain.
Can calculate a new grid that is sparse (if long_to_short = TRUE) or denser
(if long_to_short = FALSE)
}
