% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{inverse_warps}
\alias{inverse_warps}
\title{Calculate inverse warping functions}
\usage{
inverse_warps(data)
}
\arguments{
\item{data}{A dataframe with columns \code{intensity} and \code{gam}.}
}
\description{
Given a dataset with columns \code{intensity} and \code{gam}  will
calculate the inverse warping functions
}
