% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_mean.R
\name{map_to_mean}
\alias{map_to_mean}
\title{create labeled dataframe of vectorized voxel intensities for a list of nifti objects}
\usage{
map_to_mean(filepaths, ids, intensity_maximum, rescale_intensities = FALSE,
  white_stripe = FALSE, grid_length = 100, ...)
}
\arguments{
\item{filepaths}{List of paths to where nifti objects are stored.}

\item{ids}{List of unique identifiers for each image.}

\item{intensity_maximum}{Maximum value of intensity for creating grid over which to evaluate CDF and
estimate warping functions. Needs to be chosen based on inspection of data.}

\item{rescale_intensities}{If \code{TRUE}, intensities will be rescaled by their 99.9% quantile.}

\item{white_stripe}{If \code{TRUE} image will be white stripe normalized
before it is vectorized.}

\item{grid_length}{Length of downsampled CDFs to be aligned via \code{fdasrvf::time_warping()}}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame of vectorized images with columns \code{intensity}, \code{id},
\code{site}, and \code{scan}, \code{cdf}.
}
\description{
Function used along with \code{mica::vectorize_image()} to vectorize images and place in a
dataframe, where names for each image are drawn from the filepath of the
image.
}
\examples{
\dontrun{
filenames = list.files(path = "../Documents/", pattern = "flair_reg", recursive = TRUE)
filenames = paste0("../Documents/", filenames)
sitenames = lapply(filenames, function(f) substring(f, 64, 66))
scan_nos = lapply(filenames, function(f) substring(str_split(f, "Scan_")[[1]][2], 1, 1))
ids = paste(sitenames, scan_nos, sep = "_")

intensity_data = make_intensities_df(filenames, ids, sitenames, scan_nos)
}

}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
