% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_scanner.R
\name{map_to_scanner}
\alias{map_to_scanner}
\title{Map NIfTIs from one scanner to another}
\usage{
map_to_scanner(inpaths, outpath, ids, scanner, map_from, map_to,
  white_stripe = FALSE, type = NULL, grid_length = 100, ...)
}
\arguments{
\item{inpaths}{List of paths to where NIfTI objects are stored.}

\item{outpath}{Character, directory where mica normalized NIfTI objects will be stored.}

\item{ids}{List of unique identifiers for each subject. Should be character vector same length as inpaths.}

\item{scanner}{Scanner on which each image was collected. Should be character vector same length as inpaths.}

\item{map_from}{Character, name of scanner from which images will be mapped.}

\item{map_to}{Character, name of scanner to which images will be mapped.}

\item{white_stripe}{If \code{TRUE} image will be white stripe normalized
before it is vectorized.}

\item{type}{If white_stripe = TRUE, user must specify the type of image, from options
\code{type = c("T1", "T2", "FA", "MD", "first", "last", "largest")}.}

\item{grid_length}{Numeric, length of downsampled CDFs to be aligned via \code{fdasrvf::time_warping()}}

\item{...}{Additional arguments passed to or from other functions.}
}
\description{
This function reads in filepaths for a set of NIfTIs from two scanners and maps NIfTIs from one scanner to the
other.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
