% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_function.R
\name{MCC}
\alias{MCC}
\title{Multi-class Correlation (MCC)}
\usage{
MCC(study1.data, study1.label, study2.data, study2.label, w.est = NULL)
}
\arguments{
\item{study1.data}{a data frame or matrix with rows for the samples and columns for the genes. (study 1)}

\item{study1.label}{a vector of the class labels for study 1.}

\item{study2.data}{a data frame or matrix with rows for the samples and columns for the genes. (study 2)}

\item{study2.label}{a vector of the class labels for study 1.}

\item{w.est}{a vector of pre-defined class weights that add up to 1. If not specified, it will be calculated from the data.}
}
\value{
multi-class correlation (MCC) between study 1 and study 2.
}
\description{
This function is used to calculate the multi-class correlation (MCC) between two studies of gene expression, where the gene expressions are assumed to follow Gaussian distribution.
}
\references{
Lu, Shuya, Jia Li, Chi Song, Kui Shen, and George C. Tseng. 2010. “Biomarker Detection in the Integration of Multiple Multi-Class Genomic Studies.” Bioinformatics  26 (3): 333–40. https://doi.org/10.1093/bioinformatics/btp669.
}
