% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_description.R
\docType{data}
\name{metabolism_label}
\alias{metabolism_label}
\title{Mouse metabolism label list}
\format{
A list of vectors (wt, VLCAD, LCAD) for the genotypes of the samples.
}
\usage{
metabolism_label
}
\description{
A list of vectors for the labels of \link{metabolism_data}.
}
\references{
Lu, S., Li, J., Song, C., Shen, K., & Tseng, G. C. (2010). Biomarker detection in the integration of multiple multi-class genomic studies. Bioinformatics, 26(3), 333-340.
}
\keyword{datasets}
