% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_main_function.R
\name{mica.full}
\alias{mica.full}
\title{Full function for Mutual Information Concordance Analysis (MICA)}
\usage{
mica.full(
  study.data.matrix.list,
  study.label.list,
  n.perm = 500,
  p.threshold = 0.05,
  n.parallel = 1,
  post.hoc = TRUE
)
}
\arguments{
\item{study.data.matrix.list}{list of matrices for gene expression with genes in rows and samples in columns}

\item{study.label.list}{list of vectors for sample labels}

\item{n.perm}{number of permutations}

\item{p.threshold}{p-value threshold for post-hoc analysis}

\item{n.parallel}{number of parallel cores}

\item{post.hoc}{logical for post-hoc analysis}
}
\value{
A list with the following components:
 \item{gmi.stat}{statistics for generalized one-sided corrected mutual information (gMI+)}
 \item{minmcc.stat}{statistics for min-MCC}
 \item{pval.gmi}{p-values for gMI+}
 \item{pval.minmcc}{p-values for min-mcc}
 \item{pairwise.tbl}{table for post-hoc pairwise comparsion}
}
\description{
`mica.full()` is the main function for MICA, which runs the MICA and min-MCC simultaneously.
MICA first detects biomarkers with concordant multi-class patterns across partial or all of the omics studies using a global test by mutual information.
A post hoc analysis is then performed for each detected biomarkers and identify studies with concordant pattern.
}
