% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_function.R
\name{mscc.permute.test}
\alias{mscc.permute.test}
\title{Permutation Test for min-MCC, MC-TC, and Post-hoc MC-MI Analysis}
\usage{
mscc.permute.test(
  study.data.list,
  study.label.list,
  w.est = NULL,
  n.perm = 1000,
  n.parallel = 50
)
}
\arguments{
\item{study.data.list}{a list of data frames or matrices with rows for the samples and columns for the genes.}

\item{study.label.list}{a list of vectors of the class labels.}

\item{w.est}{a vector of pre-defined class weights that add up to 1. If not specified, it will be calculated from the data.}

\item{n.perm}{permutation times.}

\item{n.parallel}{number of cores used for parallel computation.}
}
\value{

}
\description{
Permutation Test for min-MCC, MC-TC, and Post-hoc MC-MI Analysis
}
