% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{upsample_hinv}
\alias{upsample_hinv}
\title{Upsample inverse warping functions}
\usage{
upsample_hinv(long_data, short_data)
}
\arguments{
\item{long_data}{A dataframe with columns of nifti-length CDFs and intensity values.}

\item{short_data}{A downsampled dataframe with columns \code{intensity} and \code{h_inv}.}
}
\description{
Given a dataset with columns \code{intensity} and \code{gam}  will
upsample the inverse warping functions
}
