% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_image.R
\name{vectorize_image}
\alias{vectorize_image}
\title{vectorize and name nifti object}
\usage{
vectorize_image(filepath, site_scan = NULL, white_stripe = FALSE, ...)
}
\arguments{
\item{filepath}{Path to where nifti object is stored.}

\item{site_scan}{Character valued ID for site and scan number in the format site_scannumber.}

\item{white_stripe}{If \code{TRUE} image will be white stripe normalized
before it is vectorized.}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame with a single vectorized image.
}
\description{
Function used to read in nifti object, vectorize, and place in a
dataframe. Option to whitestripe normalize image if desired.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
