% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cdf.R
\name{estimate_cdf}
\alias{estimate_cdf}
\title{Calulate eCDF and downsample to regular grid.}
\usage{
estimate_cdf(intensity_df, grid_length = 1000, ...)
}
\arguments{
\item{intensity_df}{Dataframe of vectorized intensity values.}

\item{grid_length}{Length of downsampled CDFs to be aligned}

\item{...}{Additional arguments passed to or from other functions.}
}
\description{
Function used to calculate empirical CDF of images and downsample to a
grid of regularly-spaced intensity values.
}
