% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cdf.R
\name{estimate_cdf}
\alias{estimate_cdf}
\title{Calulate eCDF and downsample to regular grid.}
\usage{
<<<<<<< HEAD
estimate_cdf(intensity_df, intensity_maximum = NULL,
  rescale_intensities = FALSE, grid_length = 100, ...)
=======
estimate_cdf(intensity_df, intensity_maximum,
  rescale_intensities = FALSE, white_stripe = FALSE, type = NULL,
  grid_length = 100, ...)
>>>>>>> 22e24694a8abe9a228eecfcd9710b69fd24cd67a
}
\arguments{
\item{intensity_df}{Dataframe of vectorized intensity values.}

\item{intensity_maximum}{Maximum value of intensity for creating grid over which to evaluate CDF and
estimate warping functions. Needs to be chosen based on inspection of data.}

\item{rescale_intensities}{If \code{TRUE}, intensities will be rescaled by their 99.9% quantile.
Defaults to \code{FALSE}.}

\item{grid_length}{Length of downsampled CDFs to be aligned via \code{fdasrvf::time_warping()}}

\item{...}{Additional arguments passed to or from other functions.}

\item{white_stripe}{If \code{TRUE} image will be white stripe normalized}

\item{type}{If \code{white_stripe = TRUE}, user must specify the type of image, from options
\code{type = c("T1", "T2", "FA", "MD", "first", "last", "largest")}.}
}
\description{
Function used to calculate empirical CDF of images and downsample to a
grid of regularly-spaced intensity values.
}
