% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_intensity_df.R
\name{make_intensity_df}
\alias{make_intensity_df}
\title{create labeled dataframe of vectorized voxel intensities for a list of nifti objects}
\usage{
make_intensity_df(filepaths, subj_scan_scanner, white_striped, ...)
}
\arguments{
\item{filepaths}{List of paths to where nifti objects are stored.}

\item{subj_scan_scanner}{Character valued ID for subject, scan number, and scanner, in the format
subj_scan_scanner.}

\item{white_striped}{Has the data been intensity normalized using White Stripe? Defaults to FALSE.}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame of vectorized images with columns \code{intensity}, \code{id},
\code{site}, and \code{scan}, \code{cdf}.
}
\description{
Function used along with \code{mica::vectorize_image()} to vectorize images and place in a
dataframe, where names for each image are drawn from the filepath of the
image.
}
\examples{
\dontrun{
filenames = list.files(path = "../Documents/", pattern = "flair_reg", recursive = TRUE)
filenames = paste0("../Documents/", filenames)
sitenames = lapply(filenames, function(f) substring(f, 64, 66))
scan_nos = lapply(filenames, function(f) substring(str_split(f, "Scan_")[[1]][2], 1, 1))
ids = paste(sitenames, scan_nos, sep = "_")

intensity_data = make_intensities_df(filenames, ids, sitenames, scan_nos)
}

}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
