% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_intensity_df.R
\name{make_intensity_df}
\alias{make_intensity_df}
\title{create labeled dataframe of vectorized voxel intensities for a list of nifti objects}
\usage{
make_intensity_df(filepaths, ids, sitenames = NULL, scan_nos = NULL,
  white_stripe = FALSE, type = NULL, ...)
}
\arguments{
\item{filepaths}{List of paths to where nifti objects are stored.}

\item{ids}{List of unique identifiers for each image.}

\item{sitenames}{List of character valued IDs for imaging site.}

\item{scan_nos}{List of character valued IDs for scan number. If NULL, scan number will
be assumed to be 1.}

\item{white_stripe}{If \code{TRUE} image will be white stripe normalized
before it is vectorized}

\item{type}{If \code{white_stripe = TRUE}, user must specify the type of image, from options
\code{type = c("T1", "T2", "FA", "MD", "first", "last", "largest")}.}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame of vectorized images with columns \code{intensity}, \code{id},
\code{site}, and \code{scan}, \code{cdf}.
}
\description{
Function used along with \code{mica::vectorize_image()} to vectorize images and place in a
dataframe, where names for each image are drawn from the filepath of the
image.
}
\examples{
\dontrun{
filenames = list.files(path = "../Documents/", pattern = "flair_reg", recursive = TRUE)
filenames = paste0("../Documents/", filenames)
sitenames = lapply(filenames, function(f) substring(f, 64, 66))
scan_nos = lapply(filenames, function(f) substring(str_split(f, "Scan_")[[1]][2], 1, 1))
ids = paste(sitenames, scan_nos, sep = "_")

intensity_data = make_intensities_df(filenames, ids, sitenames, scan_nos)
}

}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
