% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_mean.R
\name{map_to_mean}
\alias{map_to_mean}
\title{Normalize a set of niftis to their Karcher mean.}
\usage{
map_to_mean(
  inpaths,
  outpath,
  ids,
  intensity_maximum,
  rescale_intensities = FALSE,
  grid_length = 100,
  ...
)
}
\arguments{
\item{inpaths}{List of paths to where nifti objects are stored.}

\item{outpath}{Directory where mica normalized nifti objects will be stored}

\item{ids}{List of unique identifiers for each image.}

\item{intensity_maximum}{Maximum value of intensity for creating grid over which to evaluate CDF and
estimate warping functions. Needs to be chosen based on inspection of data.}

\item{rescale_intensities}{If \code{TRUE}, intensities will be rescaled by their 99.9 percent quantile.}

\item{grid_length}{Length of downsampled CDFs to be aligned via \code{fdasrvf::time_warping()}}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame of vectorized images with columns \code{intensity}, \code{id},
\code{site}, and \code{scan}, \code{cdf}.
}
\description{
This function reads in filepaths for a set of niftis and outputs mica-normalized niftis to a specified folder.
Also returns dataframes of vectorized niftis and warping functions.
}
\examples{
\dontrun{
filenames = list.files(path = "../Documents/", pattern = "flair_reg", recursive = TRUE)
filenames = paste0("../Documents/", filenames)
sitenames = lapply(filenames, function(f) substring(f, 64, 66))
scan_nos = lapply(filenames, function(f) substring(str_split(f, "Scan_")[[1]][2], 1, 1))
ids = paste(sitenames, scan_nos, sep = "_")

intensity_data = make_intensities_df(filenames, ids, sitenames, scan_nos)
}

}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
