% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_scanner.R
\name{map_to_scanner}
\alias{map_to_scanner}
\title{Map niftis from one scanner to another}
\usage{
map_to_scanner(
  inpaths,
  outpath,
  ids,
  scanner,
  intensity_maximum = NULL,
  map_from,
  map_to,
  white_stripe = FALSE,
  type = NULL,
  grid_length = 100,
  ...
)
}
\arguments{
\item{inpaths}{List of paths to where nifti objects are stored.}

\item{outpath}{Directory where mica normalized nifti objects will be stored.}

\item{ids}{List of unique identifiers for each subject.}

\item{scanner}{Scanner on which each image was collected. Should be character vector same length as ids.}

\item{intensity_maximum}{Maximum value of intensity for creating grid over which to evaluate CDF and
estimate warping functions. Needs to be chosen based on inspection of data.}

\item{map_from}{Character, name of scanner from which images will be mapped.}

\item{map_to}{Character, name of scanner to which images will be mapped.}

\item{white_stripe}{If \code{TRUE} image will be white stripe normalized
before it is vectorized.}

\item{type}{If white_stripe = TRUE, user must specify the type of image, from options
\code{type = c("T1", "T2", "FA", "MD", "first", "last", "largest")}.}

\item{grid_length}{Length of downsampled CDFs to be aligned via \code{fdasrvf::time_warping()}}

\item{...}{Additional arguments passed to or from other functions.}
}
\description{
This function reads in filepaths for a set of niftis from two scanners and maps niftis from one scanner to the
other.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
