% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_image.R
\name{vectorize_image}
\alias{vectorize_image}
\title{Vectorize and name nifti object}
\usage{
vectorize_image(
  filepath,
  site_scan_id = NULL,
  white_stripe = FALSE,
  type = NULL,
  ...
)
}
\arguments{
\item{filepath}{Path to where nifti object is stored.}

\item{site_scan_id}{Character valued ID for site, scan number, and id in the format
site_scannumber_id.}

\item{white_stripe}{If \code{TRUE} image will be white stripe normalized
before it is vectorized.}

\item{type}{If \code{white_stripe = TRUE}, user must specify the type of image, from options
\code{type = c("T1", "T2", "FA", "MD", "first", "last", "largest")}.}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame with a single vectorized image.
}
\description{
Function used to read in nifti object, vectorize, and place in a
dataframe. Option to whitestripe normalize image if desired.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
