% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_image.R
\name{vectorize_image}
\alias{vectorize_image}
\title{Vectorize and name nifti object}
\usage{
vectorize_image(filepath, subj_scan_scanner, ...)
}
\arguments{
\item{filepath}{Path to where nifti object is stored.}

\item{subj_scan_scanner}{Character valued ID for subject, scan number, and scanner, in the format
subj_scan_scanner.}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame with a single vectorized image.
}
\description{
Function used to read in nifti object, vectorize, and place in a
dataframe.
}
\author{
Julia Wrobel \email{julia.wrobel@cuanschutz.edu}
}
