% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_image.R
\name{vectorize_image}
\alias{vectorize_image}
\title{Vectorize and name nifti object}
\usage{
vectorize_image(filepath, site_id = NULL, filter_skull = TRUE, ...)
}
\arguments{
\item{filepath}{Path to where nifti object is stored.}

\item{site_id}{Character valued ID for site, and id in the format
site_id.}

\item{...}{Additional arguments passed to or from other functions.}
}
\value{
a data frame with a single vectorized image.
}
\description{
Function used to read in nifti object, vectorize, and place in a
dataframe. Option to whitestripe normalize image if desired.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
